+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/pyaud/compare/v3.13.2...HEAD)
 ------------------------------------------------------------------------
+### Fixed
+- Fixed issue where single files were not updating once hashed
 
 [3.13.2](https://github.com/jshwi/pyaud/releases/tag/v3.13.2) - 2022-04-29
 ------------------------------------------------------------------------